Refactor code / clean up
