+++ CARexplorer.py
@@ -6,6 +6,7 @@ from PIL import ImageTk,Image
 from tkinter import filedialog
 from tkinter import simpledialog
 from carread import *
+from carwrite import *
 from dict import *
 from get_data import *
 
@@ -48,6 +49,15 @@ def opencar():
 def clear_frame(framename):
     for widgets in framename.winfo_children():
         widgets.destroy()
+
+def updatemoneydisplay(balance):
+    moneycanvas = Canvas(mf_rightframe,width=280,borderwidth=0, highlightthickness=0,bg=colortone)
+    moneycanvas.place(x=10,y=130)
+
+    teamnamebox_5 = Label(moneycanvas,bg=colortone,width=15,height=1,bd=0,highlightthickness=0, font=('swos',14),anchor='w',text='BANK BALANCE:',fg='yellow')
+    teamnamebox_5.grid(row=0,column=0,columnspan=2)
+    teamnamebox_6 = Label(moneycanvas,bg=colortone,width=20,height=1,bd=0,highlightthickness=0, font=('swos',12),anchor='e',text="{:0,.2f}".format(float(balance)),fg='white')
+    teamnamebox_6.grid(row=1,column=1,columnspan=2,pady=5)
 
 def updateview(carfile_output):
     squad = carfile_output[0]
@@ -160,13 +170,7 @@ def updateview(carfile_output):
     box_br2 = Label(teamcanvas,bg=colortone,width=10,height=1,bd=0,highlightthickness=0, font=('swos',12),text=' ',fg='white')
     box_br2.grid(row=5,column=0,columnspan=2)
 
-    moneycanvas = Canvas(mf_rightframe,width=280,borderwidth=0, highlightthickness=0,bg=colortone)
-    moneycanvas.place(x=10,y=130)
-
-    teamnamebox_5 = Label(moneycanvas,bg=colortone,width=15,height=1,bd=0,highlightthickness=0, font=('swos',14),anchor='w',text='BANK BALANCE:',fg='yellow')
-    teamnamebox_5.grid(row=0,column=0,columnspan=2)
-    teamnamebox_6 = Label(moneycanvas,bg=colortone,width=20,height=1,bd=0,highlightthickness=0, font=('swos',12),anchor='e',text="{:0,.2f}".format(float(carfileteaminfo.money)),fg='white')
-    teamnamebox_6.grid(row=1,column=1,columnspan=2,pady=5)
+    updatemoneydisplay(carfileteaminfo.money)
 
 def close():
     quit()
@@ -181,73 +185,20 @@ def upd_data_teams():
 def upd_data_players():     # This feature is planned for future releases
     pass
 
-def read_hex_file(file_path):
-    # Open the file in binary read mode
-    with open(file_path, 'rb') as file:
-        hex_data = file.read()
-    return hex_data
-
-def update_hex_data(hex_data, offset, new_hex_value):
-    # Convert the hex data to a mutable bytearray
-    hex_data_array = bytearray(hex_data)
-    
-    # Update the hex value at the specified offset
-    # Assuming new_hex_value is a byte or bytes object
-    hex_data_array[offset:offset+len(new_hex_value)] = new_hex_value
-    
-    return bytes(hex_data_array)
-
-def write_hex_file(file_path, hex_data):
-    # Open the file in binary write mode
-    with open(file_path, 'wb') as file:
-        file.write(hex_data)
-
-def reverse_hex_pairs(hex_str):
-    # Reverse the hexadecimal string by pairs of 2 characters
-    return ''.join([hex_str[i:i+2] for i in range(0, len(hex_str), 2)][::-1])
-
 def change_balance():      # This feature is planned for future releases
     global pickedfile
     if pickedfile != '':
         user_input = simpledialog.askstring("Input", "Enter a new Bank Balance Amount:")
         
         try:
-            number = int(user_input)
+            newbalance = max(0, min(100000000, int(user_input))) #do not allow crazy values
         except ValueError:
             tk.messagebox.showinfo("Invalid", "Please enter a valid number.")
             return
-        
-        number = max(0, min(100000000, number))
 
-        # Convert the number to hexadecimal and remove the '0x' prefix
-        hex_value_no_prefix = hex(number)[2:]
-        
-        # Ensure the hexadecimal string is 8-length
-        hex_value_no_prefix = hex_value_no_prefix.zfill(8)
+        updateCARbalance(pickedfile, newbalance)
 
-        # Reverse the hexadecimal string by pairs of 2 characters
-        reversed_hex_value = reverse_hex_pairs(hex_value_no_prefix)
-        
-        # Convert the hexadecimal string to bytes
-        new_hex_value = bytes.fromhex(reversed_hex_value)
-
-        ba_ofst = 0xD5DC    # New bank balance (current amount of money)
-        # Read the original hex file
-        original_hex_data = read_hex_file(pickedfile)
-    
-        # Update the hex data
-        updated_hex_data = update_hex_data(original_hex_data, ba_ofst, new_hex_value)
-    
-        # Write the updated hex data back to the file
-        write_hex_file(pickedfile, updated_hex_data)
-
-        moneycanvas = Canvas(mf_rightframe,width=280,borderwidth=0, highlightthickness=0,bg=colortone)
-        moneycanvas.place(x=10,y=130)
-
-        teamnamebox_5 = Label(moneycanvas,bg=colortone,width=15,height=1,bd=0,highlightthickness=0, font=('swos',14),anchor='w',text='BANK BALANCE:',fg='yellow')
-        teamnamebox_5.grid(row=0,column=0,columnspan=2)
-        teamnamebox_6 = Label(moneycanvas,bg=colortone,width=20,height=1,bd=0,highlightthickness=0, font=('swos',12),anchor='e',text="{:0,.2f}".format(float(number)),fg='white')
-        teamnamebox_6.grid(row=1,column=1,columnspan=2,pady=5)
+        updatemoneydisplay(newbalance)
 
         tk.messagebox.showinfo("Done", "Balance updated in CAR file")
